// types/Types.ts
export interface HeaderLinksType {
  href: string,
  name: string
}
export interface BlogCardsType {
  image: string;
  comment: string;
  date: string;
  title: string;
}
 export interface ProjectCardsType {
  image: string;
  title: string;
  tech1: string;
  tech2: string;
  tech3: string;
  tech4: string;
  description: string;
  sourceCode: string;
  livePreview: string;
}
 export interface SkillCardsType {
  image: string;
  percent: string;
  title: string;
}
export interface ServicesCardsType {
  title: string;
  num: string;
  description: string;
}